Move insert_new_datamodel and report_exists to external.database.
